Moved Operation from rpc.v1.model to storage.v1.model
