+++ cosmos-json/src/main/scala/com/mesosphere/cosmos/rpc/v1/circe/Decoders.scala
@@ -2,12 +2,11 @@ package com.mesosphere.cosmos.rpc.v1.circe
 
 import cats.data.Xor
 import com.mesosphere.cosmos.rpc.v1.model._
+import com.mesosphere.cosmos.storage.v1.circe.Decoders._
 import com.mesosphere.cosmos.thirdparty.marathon.circe.Decoders._
 import com.mesosphere.universe
 import com.mesosphere.universe.common.circe.Decoders._
 import com.mesosphere.universe.v2.circe.Decoders._
-import com.mesosphere.universe.v2.model.PackageDetailsVersion
-import com.mesosphere.universe.v2.model.ReleaseVersion
 import com.mesosphere.universe.v3.circe.Decoders._
 import io.circe._
 import io.circe.generic.semiauto._
@@ -80,8 +79,6 @@ object Decoders {
 
   implicit val decodePackageCoordinate: Decoder[PackageCoordinate] =
     deriveDecoder[PackageCoordinate]
-
-  implicit val decodeOperation: Decoder[Operation] = deriveDecoder[Operation]
 
   implicit val decodeLocalPackage: Decoder[LocalPackage] = {
     val NotInstalledName = classOf[NotInstalled].getSimpleName
