+++ cosmos-json/src/main/scala/com/mesosphere/cosmos/rpc/v1/circe/Encoders.scala
@@ -5,9 +5,9 @@ import com.mesosphere.cosmos.thirdparty.marathon.circe.Encoders._
 import com.mesosphere.universe
 import com.mesosphere.universe.common.circe.Encoders._
 import com.mesosphere.universe.v2.circe.Encoders._
+import com.mesosphere.universe.v3.circe.Encoders._
 import com.mesosphere.universe.v3.circe.Encoders.{encodeImages => encodeV3Images}
 import com.mesosphere.universe.v3.circe.Encoders.{encodeLicense => encodeV3License}
-import com.mesosphere.universe.v3.circe.Encoders._
 import io.circe._
 import io.circe.generic.semiauto._
 import io.circe.syntax._
@@ -82,37 +82,6 @@ object Encoders {
 
   implicit val encodePackageCoordinate: Encoder[PackageCoordinate] =
     deriveEncoder[PackageCoordinate]
-
-  implicit val encodeOperation: Encoder[Operation] = deriveEncoder[Operation]
-
-    /* This encoder converts a LocalPackage into a JSON object. The total number of fields are
-     * enumerated below.
-     *
-     * {
-     *   "status": <String>,
-     *   "metadata": <PackageDefinition>,
-     *   "operation": <Operation>,
-     *   "error": <ErrorResponse>,
-     *   "packageCoordinate": <PackageCoordinate>
-     * }
-     *
-     * The 'status' will always be set while the rest of the properties are optional.
-     */
-  implicit val encodeLocalPackage = new Encoder[LocalPackage] {
-    final override def apply(value: LocalPackage): Json = {
-      val dataField = value.metadata.fold(
-        pc => ("packageCoordinate", pc.asJson),
-        pkg => ("metadata", pkg.asJson)
-      )
-
-      Json.obj(
-        "status" -> Json.fromString(value.getClass.getSimpleName),
-        dataField,
-        "error" -> value.error.asJson,
-        "operation" -> value.operation.asJson
-      )
-    }
-  }
 
   implicit val encodeServiceStartRequest: Encoder[ServiceStartRequest] =
     deriveEncoder[ServiceStartRequest]
