+++ cosmos-model/src/main/scala/com/mesosphere/cosmos/rpc/v1/model/LocalPackage.scala
@@ -1,5 +1,6 @@
 package com.mesosphere.cosmos.rpc.v1.model
 
+import com.mesosphere.cosmos.storage.v1.model.Operation
 import com.mesosphere.universe
 import com.mesosphere.universe.v3.syntax.PackageDefinitionOps._
 import scala.util.Either
