+++ cosmos-model/src/main/scala/com/mesosphere/cosmos/storage/v1/model/Operation.scala
@@ -1,4 +1,4 @@
-package com.mesosphere.cosmos.rpc.v1.model
+package com.mesosphere.cosmos.storage.v1.model
 
 import com.mesosphere.universe
 import java.util.UUID
