+++ cosmos-model/src/main/scala/com/mesosphere/cosmos/storage/v1/model/OperationFailure.scala
@@ -1,6 +1,5 @@
-package com.mesosphere.cosmos.storage.installqueue
+package com.mesosphere.cosmos.storage.v1.model
 
 import com.mesosphere.cosmos.rpc
-import com.mesosphere.cosmos.rpc.v1.model.Operation
 
 case class OperationFailure(operation: Operation, error: rpc.v1.model.ErrorResponse)
