+++ cosmos-model/src/main/scala/com/mesosphere/cosmos/storage/v1/model/OperationStatus.scala
@@ -1,0 +1,8 @@
+package com.mesosphere.cosmos.storage.v1.model
+
+sealed trait OperationStatus
+
+case class PendingStatus(operation: Operation, failure: Option[OperationFailure])
+  extends OperationStatus
+
+case class FailedStatus(failure: OperationFailure) extends OperationStatus
