+++ cosmos-model/src/test/scala/com/mesosphere/universe/v3/model/PackageDefinitionSpec.scala
@@ -1,7 +1,5 @@
 package com.mesosphere.universe.v3.model
 
-import com.mesosphere.Generators
-import java.nio.ByteBuffer
 import org.scalacheck.Gen
 import org.scalacheck.Gen.Choose
 import org.scalatest.FreeSpec
@@ -58,44 +56,4 @@ final class PackageDefinitionSpec extends FreeSpec with PropertyChecks {
 
 object PackageDefinitionSpec {
   val nonNegNum: Gen[Long] = Gen.sized(max => implicitly[Choose[Long]].choose(0, max.toLong))
-
-  val genPackageName: Gen[String] = {
-    val maxPackageNameLength = 64
-    Generators.maxSizedString(maxPackageNameLength, Gen.oneOf(Gen.numChar, Gen.alphaLowerChar))
-  }
-
-  val genReleaseVersion: Gen[PackageDefinition.ReleaseVersion] = for {
-    num <- Gen.posNum[Long]
-  } yield PackageDefinition.ReleaseVersion(num).get
-
-  val genVersion: Gen[PackageDefinition.Version] = for {
-    semver <- SemVerSpec.genSemVer
-  } yield PackageDefinition.Version(semver.toString)
-
-  val genV3Package: Gen[V3Package] = for {
-    name <- genPackageName
-    version <- genVersion
-    releaseVersion <- genReleaseVersion
-    maintainer <- Gen.alphaStr
-    description <- Gen.alphaStr
-  } yield V3Package(
-    name=name,
-    version=version,
-    releaseVersion=releaseVersion,
-    maintainer=maintainer,
-    description=description
-  )
-
-  val genV2Package: Gen[V2Package] = for {
-    v3package <- genV3Package
-  } yield V2Package(
-    name=v3package.name,
-    version=v3package.version,
-    releaseVersion=v3package.releaseVersion,
-    maintainer=v3package.maintainer,
-    description=v3package.description,
-    marathon=Marathon(ByteBuffer.allocate(0))
-  )
-
-  val genPackageDefinition: Gen[PackageDefinition] = Gen.oneOf(genV3Package, genV2Package)
 }
