+++ cosmos-model/src/test/scala/com/mesosphere/universe/v3/model/SemVerSpec.scala
@@ -1,7 +1,6 @@
 package com.mesosphere.universe.v3.model
 
-import com.mesosphere.Generators
-import org.scalacheck.Gen
+import com.mesosphere.Generators._
 import org.scalatest.FreeSpec
 import org.scalatest.Matchers
 import org.scalatest.prop.PropertyChecks
@@ -10,7 +9,7 @@ import scala.util.Random
 final class SemVerSpec extends FreeSpec with PropertyChecks with Matchers {
   "For all SemVer => String => SemVer" in {
 
-    forAll (SemVerSpec.genSemVer) { expected =>
+    forAll (genSemVer) { expected =>
         val string = expected.toString
         val actual = SemVer(string).get
 
@@ -39,38 +38,5 @@ final class SemVerSpec extends FreeSpec with PropertyChecks with Matchers {
     val actual = Random.shuffle(expected).sorted
 
     actual shouldBe expected
-  }
-}
-
-object SemVerSpec {
-  val genSemVer: Gen[SemVer] = {
-    val maxNumber = 999999999L
-    val maxStringLength = 10
-    val genNumbers = Gen.chooseNum(0, maxNumber)
-    val genPreReleases = for {
-      seqSize <- Gen.chooseNum(0, 3)
-      preReleases <- Gen.containerOfN[Seq, Either[String, Long]](
-        seqSize,
-        Gen.oneOf(
-          genNumbers.map(Right(_)),
-          Generators.maxSizedString(
-            maxStringLength,
-            Gen.alphaLowerChar
-          ).filter(_.nonEmpty).map(Left(_))
-        )
-      )
-    } yield preReleases
-
-    val genBuild = Generators.maxSizedString(maxStringLength, Gen.alphaLowerChar).map {
-      string => if (string.isEmpty) None else Some(string)
-    }
-
-    for {
-      major <- genNumbers
-      minor <- genNumbers
-      patch <- genNumbers
-      preReleases <- genPreReleases
-      build <- genBuild
-    } yield SemVer(major, minor, patch, preReleases, build)
   }
 }
