+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/circe/Decoders.scala
@@ -3,19 +3,8 @@ package com.mesosphere.cosmos.circe
 import cats.data.Xor
 import com.mesosphere.cosmos.CirceError
 import com.mesosphere.cosmos.model.ZooKeeperStorageEnvelope
-import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
-import com.mesosphere.cosmos.rpc.v1.model.Install
-import com.mesosphere.cosmos.rpc.v1.model.Operation
-import com.mesosphere.cosmos.rpc.v1.model.Uninstall
-import com.mesosphere.cosmos.rpc.v1.model.UniverseInstall
-import com.mesosphere.cosmos.storage.installqueue.OperationFailure
-import com.mesosphere.cosmos.storage.installqueue.PendingOperation
-import com.mesosphere.cosmos.storage.installqueue._
 import com.mesosphere.universe.common.circe.Decoders._
-import com.mesosphere.universe.v3.circe.Decoders._
 import io.circe.Decoder
-import io.circe.DecodingFailure
-import io.circe.HCursor
 import io.circe.generic.semiauto._
 
 object Decoders {
@@ -28,59 +17,5 @@ object Decoders {
 
   implicit val decodeZooKeeperStorageEnvelope: Decoder[ZooKeeperStorageEnvelope] =
     deriveDecoder[ZooKeeperStorageEnvelope]
-
-  implicit val decodeInstall: Decoder[Install] =
-    deriveDecoder[Install]
-
-  implicit val decodeUniverseInstall: Decoder[UniverseInstall] =
-    deriveDecoder[UniverseInstall]
-
-  implicit val decodeUninstall: Decoder[Uninstall] =
-    deriveDecoder[Uninstall]
-
-  implicit val decodeOperation: Decoder[Operation] = {
-    val InstallName: String = classOf[Install].getSimpleName
-    val UninstallName: String = classOf[Uninstall].getSimpleName
-    val UniverseInstallName: String = classOf[UniverseInstall].getSimpleName
-
-    Decoder.instance { (hc: HCursor) =>
-      hc.downField("type").as[String].flatMap {
-        case InstallName => hc.as[Install]
-        case UniverseInstallName => hc.as[UniverseInstall]
-        case UninstallName => hc.as[Uninstall]
-        case tp: String => Xor.left(DecodingFailure(
-          s"Encountered unknown type [$tp]" +
-            " while trying to decode Operation", hc.history))
-      }
-    }
-  }
-
-  implicit val decodeOperationFailure: Decoder[OperationFailure] =
-    deriveDecoder[OperationFailure]
-
-  implicit val decodePendingOperation: Decoder[PendingOperation] =
-    deriveDecoder[PendingOperation]
-
-  implicit val decodePending: Decoder[Pending] =
-    deriveDecoder[Pending]
-
-  implicit val decodeFailed: Decoder[Failed] =
-    deriveDecoder[Failed]
-
-  implicit val decodeOperationStatus: Decoder[OperationStatus] = {
-    val PendingName = classOf[Pending].getSimpleName
-    val FailedName = classOf[Failed].getSimpleName
-
-    Decoder.instance { (hc: HCursor) =>
-      hc.downField("type").as[String].flatMap {
-        case PendingName => hc.as[Pending]
-        case FailedName => hc.as[Failed]
-        case tp: String => Xor.left(DecodingFailure(
-          s"Encountered unknown type [$tp]" +
-            " while trying to decode OperationStatus", hc.history)
-        )
-      }
-    }
-  }
 
 }
