+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/storage/installqueue/ProducerView.scala
@@ -1,7 +1,7 @@
 package com.mesosphere.cosmos.storage.installqueue
 
 import com.mesosphere.cosmos.rpc
-import com.mesosphere.cosmos.rpc.v1.model.Operation
+import com.mesosphere.cosmos.storage.v1.model.Operation
 import com.twitter.util.Future
 
 /** This will be used by the producer to observe the state of
