+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/storage/v1/circe/MediaTypedDecoders.scala
@@ -1,11 +1,11 @@
 package com.mesosphere.cosmos.storage.v1.circe
 
-import com.mesosphere.cosmos.circe.Decoders._
 import com.mesosphere.cosmos.finch.MediaTypedDecoder
 import com.mesosphere.cosmos.http.MediaTypes
 import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
 import com.mesosphere.cosmos.rpc.v1.model.PackageRepository
-import com.mesosphere.cosmos.storage.installqueue.OperationStatus
+import com.mesosphere.cosmos.storage.v1.circe.Decoders._
+import com.mesosphere.cosmos.storage.v1.model.OperationStatus
 
 object MediaTypedDecoders {
   implicit val repositoryListDecoder: MediaTypedDecoder[List[PackageRepository]] =
