+++ cosmos-server/src/test/scala/com/mesosphere/cosmos/storage/installqueue/InstallQueueSpec.scala
@@ -5,10 +5,15 @@ import com.mesosphere.cosmos.OperationInProgress
 import com.mesosphere.cosmos.model.ZooKeeperUri
 import com.mesosphere.cosmos.rpc.v1.model.ErrorResponse
 import com.mesosphere.cosmos.rpc.v1.model.PackageCoordinate
-import com.mesosphere.cosmos.rpc.v1.model.UniverseInstall
 import com.mesosphere.cosmos.storage.Envelope
 import com.mesosphere.cosmos.storage.v1.circe.MediaTypedDecoders._
 import com.mesosphere.cosmos.storage.v1.circe.MediaTypedEncoders._
+import com.mesosphere.cosmos.storage.v1.model.FailedStatus
+import com.mesosphere.cosmos.storage.v1.model.OperationFailure
+import com.mesosphere.cosmos.storage.v1.model.OperationStatus
+import com.mesosphere.cosmos.storage.v1.model.PendingStatus
+import com.mesosphere.cosmos.storage.v1.model.PendingOperation
+import com.mesosphere.cosmos.storage.v1.model.UniverseInstall
 import com.mesosphere.cosmos.zookeeper.Clients
 import com.mesosphere.universe.test.TestingPackages
 import com.mesosphere.universe.v3.model.PackageDefinition
@@ -49,7 +54,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
         )
         assertResult(())(addResult)
 
-        checkInstallQueueContents(client, coordinate1, Pending(universeInstall, None))
+        checkInstallQueueContents(client, coordinate1, PendingStatus(universeInstall, None))
       }
 
       "when the parent path exists but the status does not" in { testParameters =>
@@ -63,12 +68,12 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
         checkInstallQueueContents(
           client,
           coordinate1,
-          Pending(universeInstall, None))
+          PendingStatus(universeInstall, None))
       }
 
       "on a coordinate that has a pending operation but no failures" in { testParameters =>
         val (client, installQueue) = testParameters
-        val pendingUniverseInstall = Pending(universeInstall, None)
+        val pendingUniverseInstall = PendingStatus(universeInstall, None)
 
         insertPackageStatusIntoQueue(
           client,
@@ -90,7 +95,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
         insertPackageStatusIntoQueue(
           client,
           coordinate1,
-          Failed(OperationFailure(universeInstall, errorResponse1)))
+          FailedStatus(OperationFailure(universeInstall, errorResponse1)))
 
         val addResult = Await.result(
           installQueue.add(coordinate1, universeInstall)
@@ -100,14 +105,14 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
         checkInstallQueueContents(
           client,
           coordinate1,
-          Pending(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
+          PendingStatus(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
       }
 
       "on a coordinate that has an operation and a failure" in { testParameters =>
         val (client, installQueue) = testParameters
 
         val pendingUniverseInstallWithFailure =
-          Pending(universeInstall, Some(OperationFailure(universeInstall, errorResponse1)))
+          PendingStatus(universeInstall, Some(OperationFailure(universeInstall, errorResponse1)))
 
         insertPackageStatusIntoQueue(
           client,
@@ -138,8 +143,8 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
       assertResult(())(add1)
       assertResult(())(add2)
 
-      checkInstallQueueContents(client, coordinate1, Pending(universeInstall, None))
-      checkInstallQueueContents(client, coordinate2, Pending(universeInstall, None))
+      checkInstallQueueContents(client, coordinate1, PendingStatus(universeInstall, None))
+      checkInstallQueueContents(client, coordinate2, PendingStatus(universeInstall, None))
     }
   }
 
@@ -169,14 +174,14 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
 
         "on a coordinate that has a pending operation but no failures" in { testParameters =>
           val (client, installQueue) = testParameters
-          insertPackageStatusIntoQueue(client, coordinate1, Pending(universeInstall, None))
+          insertPackageStatusIntoQueue(client, coordinate1, PendingStatus(universeInstall, None))
           Await.result(
             installQueue.failure(coordinate1, errorResponse1)
           )
           checkInstallQueueContents(
             client,
             coordinate1,
-            Failed(OperationFailure(universeInstall, errorResponse1)))
+            FailedStatus(OperationFailure(universeInstall, errorResponse1)))
         }
 
         "on a coordinate that has a failed operation, but no pending operation" in { testParameters =>
@@ -184,7 +189,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Failed(OperationFailure(universeInstall, errorResponse1)))
+            FailedStatus(OperationFailure(universeInstall, errorResponse1)))
           val error = intercept[InstallQueueError](
             Await.result(
               installQueue.failure(coordinate1, errorResponse1)
@@ -198,21 +203,21 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Pending(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
+            PendingStatus(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
           Await.result(
             installQueue.failure(coordinate1, errorResponse2)
           )
           checkInstallQueueContents(
             client,
             coordinate1,
-            Failed(OperationFailure(universeInstall, errorResponse2)))
+            FailedStatus(OperationFailure(universeInstall, errorResponse2)))
         }
       }
 
       "Fail multiple non-conflicting operations" in { testParameters =>
         val (client, installQueue) = testParameters
-        insertPackageStatusIntoQueue(client, coordinate1, Pending(universeInstall, None))
-        insertPackageStatusIntoQueue(client, coordinate2, Pending(universeInstall, None))
+        insertPackageStatusIntoQueue(client, coordinate1, PendingStatus(universeInstall, None))
+        insertPackageStatusIntoQueue(client, coordinate2, PendingStatus(universeInstall, None))
         val failTwoOperations =
           for {
             _ <- installQueue.failure(coordinate1, errorResponse1)
@@ -222,11 +227,11 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
         checkInstallQueueContents(
           client,
           coordinate1,
-          Failed(OperationFailure(universeInstall, errorResponse1)))
+          FailedStatus(OperationFailure(universeInstall, errorResponse1)))
         checkInstallQueueContents(
           client,
           coordinate2,
-          Failed(OperationFailure(universeInstall, errorResponse1)))
+          FailedStatus(OperationFailure(universeInstall, errorResponse1)))
       }
     }
 
@@ -255,7 +260,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
 
         "on a coordinate that has a pending operation but no failures" in { testParameters =>
           val (client, installQueue) = testParameters
-          insertPackageStatusIntoQueue(client, coordinate1, Pending(universeInstall, None))
+          insertPackageStatusIntoQueue(client, coordinate1, PendingStatus(universeInstall, None))
           Await.result(
             installQueue.success(coordinate1)
           )
@@ -267,7 +272,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Failed(OperationFailure(universeInstall, errorResponse1)))
+            FailedStatus(OperationFailure(universeInstall, errorResponse1)))
           val error = intercept[InstallQueueError](
             Await.result(
               installQueue.success(coordinate1)
@@ -281,7 +286,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Pending(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
+            PendingStatus(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
           Await.result(
             installQueue.success(coordinate1)
           )
@@ -291,8 +296,8 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
 
       "Success on multiple non-conflicting operations" in { testParameters =>
         val (client, installQueue) = testParameters
-        insertPackageStatusIntoQueue(client, coordinate1, Pending(universeInstall, None))
-        insertPackageStatusIntoQueue(client, coordinate2, Pending(universeInstall, None))
+        insertPackageStatusIntoQueue(client, coordinate1, PendingStatus(universeInstall, None))
+        insertPackageStatusIntoQueue(client, coordinate2, PendingStatus(universeInstall, None))
         val successOnTwoOperations =
           for {
             _ <- installQueue.success(coordinate1)
@@ -328,7 +333,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Pending(universeInstall, None))
+            PendingStatus(universeInstall, None))
           val nextPendingOperation = Await.result(
             installQueue.next()
           )
@@ -342,7 +347,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Failed(OperationFailure(universeInstall, errorResponse1)))
+            FailedStatus(OperationFailure(universeInstall, errorResponse1)))
           val nextPendingOperation = Await.result(
             installQueue.next()
           )
@@ -354,7 +359,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Pending(
+            PendingStatus(
               universeInstall,
               Some(OperationFailure(
                 universeInstall, errorResponse1))))
@@ -375,23 +380,23 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           insertPackageStatusIntoQueue(
             client,
             coordinate1,
-            Pending(universeInstall, None))
+            PendingStatus(universeInstall, None))
           insertPackageStatusIntoQueue(
             client,
             coordinate2,
-            Pending(universeInstall, None))
+            PendingStatus(universeInstall, None))
           insertPackageStatusIntoQueue(
             client,
             coordinate3,
-            Failed(OperationFailure(universeInstall, errorResponse1)))
+            FailedStatus(OperationFailure(universeInstall, errorResponse1)))
           insertPackageStatusIntoQueue(
             client,
             coordinate4,
-            Failed(OperationFailure(universeInstall, errorResponse2)))
+            FailedStatus(OperationFailure(universeInstall, errorResponse2)))
           insertPackageStatusIntoQueue(
             client,
             coordinate5,
-            Pending(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
+            PendingStatus(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
 
           val n1 = Await.result(installQueue.next())
           removePackageStatusFromQueue(client, coordinate1)
@@ -427,23 +432,23 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
         insertPackageStatusIntoQueue(
           client,
           coordinate1,
-          Pending(universeInstall, None))
+          PendingStatus(universeInstall, None))
         insertPackageStatusIntoQueue(
           client,
           coordinate2,
-          Pending(universeInstall, None))
+          PendingStatus(universeInstall, None))
         insertPackageStatusIntoQueue(
           client,
           coordinate3,
-          Failed(OperationFailure(universeInstall, errorResponse1)))
+          FailedStatus(OperationFailure(universeInstall, errorResponse1)))
         insertPackageStatusIntoQueue(
           client,
           coordinate4,
-          Failed(OperationFailure(universeInstall, errorResponse1)))
+          FailedStatus(OperationFailure(universeInstall, errorResponse1)))
         insertPackageStatusIntoQueue(
           client,
           coordinate5,
-          Pending(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
+          PendingStatus(universeInstall, Some(OperationFailure(universeInstall, errorResponse1))))
 
         val callNextTwice =
           for {
@@ -452,7 +457,7 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
           } yield (n1, n2)
         val (n1, n2) = Await.result(callNextTwice)
 
-        checkInstallQueueContents(client, coordinate1, Pending(universeInstall, None))
+        checkInstallQueueContents(client, coordinate1, PendingStatus(universeInstall, None))
         removePackageStatusFromQueue(client, coordinate1)
 
         val callNextTwiceAgain =
@@ -479,11 +484,11 @@ final class InstallQueueSpec extends fixture.FreeSpec with BeforeAndAfterAll {
       "return all pending and failed operations in the queue" in { testParameters =>
       val (client, installQueue) = testParameters
       val expectedState = Map(
-        coordinate1 -> Pending(universeInstall, None),
-        coordinate2 -> Pending(universeInstall, None),
-        coordinate3 -> Failed(OperationFailure(universeInstall, errorResponse1)),
-        coordinate4 -> Failed(OperationFailure(universeInstall, errorResponse1)),
-        coordinate5 -> Pending(universeInstall, Some(OperationFailure(universeInstall, errorResponse1)))
+        coordinate1 -> PendingStatus(universeInstall, None),
+        coordinate2 -> PendingStatus(universeInstall, None),
+        coordinate3 -> FailedStatus(OperationFailure(universeInstall, errorResponse1)),
+        coordinate4 -> FailedStatus(OperationFailure(universeInstall, errorResponse1)),
+        coordinate5 -> PendingStatus(universeInstall, Some(OperationFailure(universeInstall, errorResponse1)))
       )
 
       expectedState.foreach { case (coordinate, status) =>
